Fix compile: Float vs Double di getVisualBounds TextBox
kotlin.math.abs(cos/sin) mengembalikan Double sehingga argumen RectF
jadi Double. Konversi ke Float agar cocok dengan RectF(Float).

+++ app/src/main/java/com/grooxtyper/app/model/TextBox.kt
@@ -404,8 +404,8 @@ class TextBox(
         val hw = w / 2f + pad
         val hh = h / 2f + pad
         val rad = Math.toRadians(rotation.toDouble())
-        val c = kotlin.math.abs(cos(rad))
-        val s = kotlin.math.abs(sin(rad))
+        val c = kotlin.math.abs(cos(rad)).toFloat()
+        val s = kotlin.math.abs(sin(rad)).toFloat()
         val ew = hw * c + hh * s
         val eh = hw * s + hh * c
         return RectF(
